+++ FSharp.Json/Core.fs
@@ -107,14 +107,30 @@ module internal Core =
                 let t, value = transformToTargetType t value jsonField.Transform
                 let t = getUntypedType t value
                 match t with
+                | t when t = typeof<uint16> ->
+                    JsonValue.Number (decimal (value :?> uint16))
+                | t when t = typeof<int16> ->
+                    JsonValue.Number (decimal (value :?> int16))
                 | t when t = typeof<int> ->
                     JsonValue.Number (decimal (value :?> int))
+                | t when t = typeof<uint32> ->
+                    JsonValue.Number (decimal (value :?> uint32))
                 | t when t = typeof<int64> ->
                     JsonValue.Number (decimal (value :?> int64))
+                | t when t = typeof<uint64> ->
+                    JsonValue.Number (decimal (value :?> uint64))
+                | t when t = typeof<bigint> ->
+                    JsonValue.Number (decimal (value :?> bigint))
+                | t when t = typeof<single> ->
+                    JsonValue.Float (float (value :?> single))
                 | t when t = typeof<float> ->
                     JsonValue.Float (value :?> float)
                 | t when t = typeof<decimal> ->
                     JsonValue.Number (value :?> decimal)
+                | t when t = typeof<byte> ->
+                    JsonValue.Number (decimal (value :?> byte))
+                | t when t = typeof<sbyte> ->
+                    JsonValue.Number (decimal (value :?> sbyte))
                 | t when t = typeof<bool> ->
                     JsonValue.Boolean (value :?> bool)
                 | t when t = typeof<string> ->
@@ -305,14 +321,30 @@ module internal Core =
                 let t = getUntypedType path t jvalue
                 let jvalue =
                     match t with
+                    | t when t = typeof<int16> ->
+                        JsonValueHelpers.getInt16 path jvalue :> obj
+                    | t when t = typeof<uint16> ->
+                        JsonValueHelpers.getUInt16 path jvalue :> obj
                     | t when t = typeof<int> ->
                         JsonValueHelpers.getInt path jvalue :> obj
+                    | t when t = typeof<uint32> ->
+                        JsonValueHelpers.getUInt32 path jvalue :> obj
                     | t when t = typeof<int64> ->
                         JsonValueHelpers.getInt64 path jvalue :> obj
+                    | t when t = typeof<uint64> ->
+                        JsonValueHelpers.getUInt64 path jvalue :> obj
+                    | t when t = typeof<bigint> ->
+                        JsonValueHelpers.getBigint path jvalue :> obj
+                    | t when t = typeof<single> ->
+                        JsonValueHelpers.getSingle path jvalue :> obj
                     | t when t = typeof<float> ->
                         JsonValueHelpers.getFloat path jvalue :> obj
                     | t when t = typeof<decimal> ->
                         JsonValueHelpers.getDecimal path jvalue :> obj
+                    | t when t = typeof<byte> ->
+                        JsonValueHelpers.getByte path jvalue :> obj
+                    | t when t = typeof<sbyte> ->
+                        JsonValueHelpers.getSByte path jvalue :> obj
                     | t when t = typeof<bool> ->
                         JsonValueHelpers.getBool path jvalue :> obj
                     | t when t = typeof<string> ->