+++ FSharp.Json/JsonValueHelpers.fs
@@ -8,18 +8,54 @@ module internal JsonValueHelpers =
     let raiseWrongType path typeName jvalue =
         raise(JsonDeserializationError(path, sprintf "Expected type %s is incompatible with jvalue: %A" typeName jvalue))
 
+    let getInt16 (path: JsonPath) (jvalue: JsonValue) =
+        match jvalue with
+        | JsonValue.Number value -> int16 value
+        | JsonValue.Float value -> int16 value
+        | _ -> raiseWrongType path "int16" jvalue
+
+    let getUInt16 (path: JsonPath) (jvalue: JsonValue) =
+        match jvalue with
+        | JsonValue.Number value -> uint16 value
+        | JsonValue.Float value -> uint16 value
+        | _ -> raiseWrongType path "uint16" jvalue
+        
     let getInt (path: JsonPath) (jvalue: JsonValue) =
         match jvalue with
         | JsonValue.Number value -> int value
         | JsonValue.Float value -> int value
         | _ -> raiseWrongType path "int" jvalue
 
+    let getUInt32 (path: JsonPath) (jvalue: JsonValue) =
+        match jvalue with
+        | JsonValue.Number value -> uint32 value
+        | JsonValue.Float value -> uint32 value
+        | _ -> raiseWrongType path "uint32" jvalue
+    
     let getInt64 (path: JsonPath) (jvalue: JsonValue) =
         match jvalue with
         | JsonValue.Number value -> int64 value
         | JsonValue.Float value -> int64 value
         | _ -> raiseWrongType path "int64" jvalue
 
+    let getUInt64 (path: JsonPath) (jvalue: JsonValue) =
+        match jvalue with
+        | JsonValue.Number value -> uint64 value
+        | JsonValue.Float value -> uint64 value
+        | _ -> raiseWrongType path "uint64" jvalue
+
+    let getBigint (path: JsonPath) (jvalue: JsonValue) =
+        match jvalue with
+        | JsonValue.Number value -> bigint value
+        | JsonValue.Float value -> bigint value
+        | _ -> raiseWrongType path "bigint" jvalue
+
+    let getSingle (path: JsonPath) (jvalue: JsonValue) =
+        match jvalue with
+        | JsonValue.Float value -> single value
+        | JsonValue.Number value -> single value
+        | _ -> raiseWrongType path "single" jvalue
+    
     let getFloat (path: JsonPath) (jvalue: JsonValue) =
         match jvalue with
         | JsonValue.Float value -> value
@@ -38,6 +74,12 @@ module internal JsonValueHelpers =
         | JsonValue.Float value -> byte value
         | _ -> raiseWrongType path "byte" jvalue
 
+    let getSByte (path: JsonPath) (jvalue: JsonValue) =
+        match jvalue with
+        | JsonValue.Number value -> sbyte value
+        | JsonValue.Float value -> sbyte value
+        | _ -> raiseWrongType path "sbyte" jvalue
+    
     let getBool (path: JsonPath) (jvalue: JsonValue) =
         match jvalue with
         | JsonValue.Boolean value -> value